refactor: consolidate org-chart getNodeLabel + budget dimension switch
Extract getNodeLabel + node-data shape guards from useOrgChartSelection
into a shared web/src/pages/org/node-utils.ts. OrgChartFilter previously
had a similar but cast-based copy; now both call the guard-protected
version, so a record with a blank/whitespace name can never surface as
an empty UI label.

Replace the two parallel switch statements in computeCostBreakdown with
a DIMENSION_RESOLVERS strategy table keyed by BreakdownDimension, so the
key/label resolution stays exhaustively typed against the union and
each dimension's logic lives in one place.

Refs #1733 (R8)
