+++ web/src/pages/org/useOrgChartSelection.ts
@@ -3,7 +3,7 @@ import type { MouseEvent as ReactMouseEvent } from 'react'
 import type { Node } from '@xyflow/react'
 import { useNavigate } from 'react-router'
 import { useToastStore } from '@/stores/toast'
-import type { AgentNodeData, DepartmentGroupData, OwnerNodeData } from './build-org-tree'
+import { getNodeLabel } from './node-utils'
 
 type NodeType = ContextMenuState['nodeType']
 
@@ -15,68 +15,6 @@ const VALID_NODE_TYPES: ReadonlySet<NodeType> = new Set<NodeType>([
 
 function isValidNodeType(value: string | undefined): value is NodeType {
   return value !== undefined && (VALID_NODE_TYPES as ReadonlySet<string>).has(value)
-}
-
-/** Non-null, non-array, object-shaped value (shared pre-check for shape guards). */
-function isObjectRecord(data: unknown): data is Record<string, unknown> {
-  return typeof data === 'object' && data !== null && !Array.isArray(data)
-}
-
-/**
- * Build a type predicate that asserts every listed field is a non-empty
- * ``string`` on the input object. Used by the node-data shape guards so
- * each one validates the full string-typed surface of its interface,
- * not just one field.
- */
-function makeStringFieldGuard<T>(
-  requiredStringFields: readonly (keyof T & string)[],
-): (data: unknown) => data is T {
-  return (data: unknown): data is T => {
-    if (!isObjectRecord(data)) return false
-    for (const key of requiredStringFields) {
-      const value = (data as Record<string, unknown>)[key]
-      // Reject whitespace-only strings as well as empty ones -- a
-      // label like ``'   '`` would otherwise pass this guard and
-      // surface as a blank name / role / id in the UI instead of
-      // falling back to ``node.id`` via the outer code path.
-      if (typeof value !== 'string' || value.trim().length === 0) return false
-    }
-    return true
-  }
-}
-
-const isAgentNodeData = makeStringFieldGuard<AgentNodeData>([
-  'agentId',
-  'name',
-  'role',
-  'department',
-  'level',
-  'runtimeStatus',
-])
-
-const isDepartmentGroupData = makeStringFieldGuard<DepartmentGroupData>([
-  'departmentName',
-  'displayName',
-])
-
-const isOwnerNodeData = makeStringFieldGuard<OwnerNodeData>([
-  'ownerId',
-  'displayName',
-  'role',
-])
-
-function getNodeLabel(node: Node): string {
-  switch (node.type) {
-    case 'agent':
-    case 'ceo':
-      return isAgentNodeData(node.data) ? node.data.name : node.id
-    case 'department':
-      return isDepartmentGroupData(node.data) ? node.data.displayName : node.id
-    case 'owner':
-      return isOwnerNodeData(node.data) ? node.data.displayName : node.id
-    default:
-      return node.id
-  }
 }
 
 export interface ContextMenuState {
